edit RoleController And Test

+++ backend/app/Http/Controllers/RoleController.php
@@ -8,7 +8,7 @@ use Illuminate\Http\Request;
 use App\Traits\HttpResponses;
 use App\Http\Requests\RoleRequest;
 use Illuminate\Support\Facades\Auth;
-
+use Illuminate\Support\Facades\DB;
 class RoleController extends Controller
 {
     use HttpResponses;
@@ -17,7 +17,7 @@ class RoleController extends Controller
      */
     public function index()
     {
-        $roles = Role::find(Auth::user()->role_id)->childrenAndSelf()->get();
+        $roles = Role::find(Auth::user()->role_id)->descendantsAndSelf()->get();
         return $this->success(data: $roles);
     }
 
@@ -27,7 +27,7 @@ class RoleController extends Controller
     public function store(Request $request)
     {
         $role_name = $request->role_name;
-        $parent_id = $request->parent_id;
+        $parent_id = Auth::user()->role_id;
         try {
             $role = Role::create([
                 'role_name' => $role_name,
@@ -41,26 +41,25 @@ class RoleController extends Controller
 
     public function show($id)
     {
-
+        $roles = Role::find(Auth::user()->role_id)->descendantsAndSelf()->where('role_id', $id)->get();
+        return $this->success(data: $roles);
     }
-    public function update(Request $request, Role $Role)
+    public function update(Request $request, $id)
     {
+        $Role = Role::find(Auth::user()->role_id);
         if (!$Role) {
-            return response()->json([
-                'status' => 2,
-                'message' => 'Role not found'
-            ], 404);
+            return $this->error(data: "ไม่สามารถ update ได้");
         } else {
-            $role_name = $request->role_name;
-            $parent_id = $request->parent_id;
-            $Role->update([
-                'role_name' => $role_name,
-                'parent_id' => $parent_id
-            ]);
-            return response()->json([
-                'status' => 1,
-                'message' => 'Update Success'
-            ], 200);
+            $roles = Role::find(Auth::user()->role_id)->descendantsAndSelf()->where('role_id', $id)->get()->count();
+            if($roles > 0){
+                $Role_update = $role_name = $request->role_name;
+                Role::where('role_id', $id)->update([
+                    'role_name' => $role_name
+                ]);
+                return $this->success(data: $Role_update);
+            }else{
+                return $this->error(data: "ไม่สามารถ update ได้");
+            }
         }
     }
 
@@ -76,11 +75,24 @@ class RoleController extends Controller
                 'message' => 'Role not found'
             ], 404);
         } else {
-            $Role->delete();
-            return response()->json([
-                'status' => 1,
-                'message' => 'Delete Success'
-            ], 200);
+            $check_descendants = Role::find(Auth::user()->role_id)->descendants()->where('role_id',$id)->get()->count();
+            // return $this->success(data: $check_descendants);
+            if($check_descendants > 0){
+                $check_mother = Role::find($id)->descendants()->get()->count();
+                // return $this->success(data: $check_mother);
+                if($check_mother == 0){
+                    $Role->delete();
+                    // $User_update = User::where("role_id",$id)->get();
+                    $User_update = User::where("role_id",$id)->update([
+                        'role_id' => null
+                    ]);
+                    return $this->success(data: $Role);
+                }else{
+                    return $this->error(data: "ไม่สามารถลบได้");
+                }
+            }else{
+                return $this->error(data: "ไม่สามารถลบได้");
+            }
         }
     }
